local room filled

+++ app/src/main/java/com/nickolay/testtask65app/data/entity/Specialty.kt
@@ -1,6 +1,6 @@
 package com.nickolay.testtask65app.data.entity
 
 data class Specialty(
-    val specialty_id: Int,
+    val specialty_id: Long,
     val name: String
 )
