+++ app/src/main/java/com/nickolay/testtask65app/data/roomdb/specialty/SpecialtyModel.kt
@@ -6,5 +6,6 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "specialty_table")
 class SpecialtyModel (
     @PrimaryKey
-    val specialtyId: Int,
-    val specialtyName: String)
+    val specialtyId: Long,
+    val specialtyName: String
+)
